+++ electron/main/index.ts
@@ -6,7 +6,7 @@ import fs from "fs"
 import { 
     WINDOW_CLOSE_EVENT, WINDOW_FULLSCREEN_STATE, WINDOW_FOCUS_STATE, SETTINGS_CHANGE_EVENT,
     TITLE_BAR_BG_LIGHT, TITLE_BAR_BG_LIGHT_BLURRED, TITLE_BAR_BG_DARK, TITLE_BAR_BG_DARK_BLURRED,
-    SCRATCH_FILE_NAME, SAVE_TABS_STATE, LOAD_TABS_STATE, CONTEXT_MENU_CLOSED,
+    SCRATCH_FILE_NAME, SAVE_TABS_STATE, LOAD_TABS_STATE, CONTEXT_MENU_CLOSED, GET_SYSTEM_LOCALE,
 } from '@/src/common/constants'
 
 import { menu, getTrayMenu, getEditorContextMenu, getTabContextMenu, getSpellcheckingContextMenu } from './menu'
@@ -556,4 +556,8 @@ ipcMain.handle(SAVE_TABS_STATE, async (event, tabsState) => {
 
 ipcMain.handle(LOAD_TABS_STATE, async (event) => {
     return CONFIG.get("openTabsState")
+})
+
+ipcMain.handle(GET_SYSTEM_LOCALE, async (event) => {
+    return app.getSystemLocale()
 })