+++ src/editor/block/block-parsing.js
@@ -142,6 +142,12 @@ export function getBlocksFromString(state) {
             };
             blocks.push(block);
         }
+        firstBlockDelimiterSize = blocks[0]?.delimiter.to
         //console.log("getBlocksFromString() took", timer(), "ms")
         return blocks
+}
+
+export function getBlockDelimiter(blockType, autoDetect, date) {
+    date = date ?? new Date()
+    return `\n∞∞∞${autoDetect ? blockType + '-a' : blockType};created=${date.toISOString()}\n`
 }