+++ tests/test-utils.js
@@ -53,6 +53,21 @@ export class HeynotePage {
         await this.page.evaluate((position) => window._heynote_editor.setCursorPosition(position), position)
     }
 
+    async getMainSelection() {
+        return await this.page.evaluate(() => window._heynote_editor.view.state.selection.main)
+    }
+
+    async expectSelectionContent(content) {
+        return await expect.poll(async () => {
+            return await this.page.evaluate(() => {
+                const state = window._heynote_editor.view.state
+                const selection = state.selection.main
+                console.log("selection:", state.doc.sliceString(selection.from, selection.to))
+                return state.doc.sliceString(selection.from, selection.to)
+            })
+        }).toBe(content)
+    }
+
     async getBlockContent(blockIndex) {
         const blocks = await this.getBlocks()
         const content = await this.getContent()
@@ -95,5 +110,13 @@ export class HeynotePage {
 
     async executeCommand(command) {
         return await this.page.evaluate((command) => window._heynote_editor.executeCommand(command), command)
+    }
+
+    async waitForContentSaved() {
+        await expect.poll(async () => {
+            const editorContent = await this.page.evaluate(() => window._heynote_editor.getContent())
+            const diskContent = await this.page.evaluate(() => window._heynote_editor.diskContent)
+            return editorContent === diskContent
+        }).toBeTruthy()
     }
 }